added context in anon-core

+++ crates/anon-parser/src/literal.rs
@@ -1,9 +1,11 @@
+#![allow(dead_code)]
+
 use crate::intern::Sym;
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub enum Literal {
     String(Sym),
     Char(char),
-    Integer(u32),
+    Integer(i64),
     Float(f64),
 }
